add new module for guiding imports of optional packages
normally, you may import something but get error about a missing package and you wonder if you didnt install the package correctly, or package is corrupted and so on. This utility enables us to exactly tell user that the SDK or API they are acessing is part of the optional dependencies and if they want to use that, they must install the package with a specific command which we tell them that too.