use `import` utility for all API codes
as we said it in #59, API does not need to be mandatory and can be optional. Package dependencies are reflect in `pyproject.toml` and now this change adds such optional behavior to code too. Now if you import SDK codes, you get no error and if you import codes belonging to optional dependencies, you get a clean error to tell you what to install.